requirements.txt
streamlit
streamlit-webrtc
tensorflow
opencv-python-headless
numpy
Pillow
